separating the schema for service cards
